Refactor : 테이블 컴포넌트 분리
- tableHead를 각각 cell 컴포넌트로 분리
