+++ src/components/tableHead/index.ts
@@ -1,0 +1,5 @@
+export { default as TimeCell } from './TimeCell';
+export { default as StatusCell } from './StatusCell';
+export { default as IdCell } from './IdCell';
+export { default as DateCell } from './DateCell';
+export { default as ClientNameCell } from './ClientNameCell';
